Run a simmr_input object with multiple groups through a simmr Markov chain Monte Carlo (MCMC) function with group as a random effect

This function allows multiple groups to be run through simmr with the group variable treated as a random effect. It takes a simmr_input object created via simmr_load, runs an MCMC to determine the proportions, and then outputs a simmr_output_re object for further analysis and plotting via summary.simmr_output_re and plot.simmr_output_re.

simmr_mcmc_group_re(
  simmr_in,
  source_control = matrix(1, nrow = simmr_in$n_groups, ncol = simmr_in$n_sources),
  scale_re = 1,
  mcmc_control = list(iter = 10000, burn = 1000, thin = 10, n.chain = 4)
)

Arguments

simmr_in
An object created via the function simmr_load. This object must have multiple groups

source_control
A binary matrix which identifies which sources are to be included in which groups. The matrix should have the same number of rows as the number of groups, and the same number of columns as the total number of sources. If NULL (default) will assume all sources contribute to the mixture for all groups

scale_re
The prior scale value of the Cauchy t-distributed random effect. For detailed prior control of individual groups use the main simmr_mcmc function.

mcmc_control
A list of values including arguments named iter (number of iterations), burn (size of burn-in), thin (amount of thinning), and n.chain (number of MCMC chains).

Value

An object of class simmr_output_re with two named top-level components:

input

The simmr_input object given to the simmr_mcmc function

output

A set of MCMC chains of class mcmc.list from the coda package. These can be analysed using the summary.simmr_output_re and plot.simmr_output_re functions.

Details

If, after running simmr_mcmc_group_re the convergence diagnostics in summary.simmr_output_re are not satisfactory, the values of iter, burn and thin in mcmc_control should be increased by a factor of 10.

References

Andrew C. Parnell, Donald L. Phillips, Stuart Bearhop, Brice X. Semmens, Eric J. Ward, Jonathan W. Moore, Andrew L. Jackson, Jonathan Grey, David J. Kelly, and Richard Inger. Bayesian stable isotope mixing models. Environmetrics, 24(6):387–399, 2013.

Andrew C Parnell, Richard Inger, Stuart Bearhop, and Andrew L Jackson. Source partitioning using stable isotopes: coping with too much variation. PLoS ONE, 5(3):5, 2010.

See also

Examples